Moved Jupiter messages to config file.
All messages used within the Jupiter Watchman were moved to the
messages.py file. The plan is to move all Watchmen messages to the
messages.py file, but there are still a lot of Watchmen to convert over.
This will be done gradually as Watchman are refactored or a developer
has the time to complete it.