Fixing Jupiter notification time issue.
Jupiter is only supposed to send SNS notifications at 8am, 12pm, or 4pm
PST. There was a bug in the code that allowed the SNS notification to be
sent at any time. This bug was caused by referencing the LATEST file in
S3 to decide if a notification was sent. The LATEST file in S3 was
supposed to reflect whether the current run was a success or not, but
Jupiter was looking at this file before updating it for the latest run,
so it was looking at the previous run's result. This messed up the logic
that decided whether or not a notification is sent in the
check_skip_notification method. Jupiter has been changed to use the
result (success or failure) of the current run to decide if an SNS
notification should be sent.

Changes in this commit include:
- Removing check_failure method, which utilized the LATEST file in S3.
- Removing the unused global variable CHECK_TIME_PDT.
- Changing check_notification_time time checking logic, and removing
  uses of is_workhour.
- Changing check_skip_notification to not use last_failed and
  simplifying the code.
- Changing the summarize method to not use the check_failure method and
  last_failed variable.