Updated Jupiter Endpoint config files
Added Sockeye test endpoints to the endpoints config files and renamed endpoints for more specific alert messages. Added a "calendar"
variable to each endpoint which will be used to determine if a
notification should be sent if the endpoint goes down. When "calendar" =
"enabled" then the endpoint will use the Infoblox calendar and the
specified notification hours to determine if a notification is sent, if
"calendar" = "disabled" then a notification will always be sent if the
endpoint goes down.