Updated Jupiter's notification configuration.
Updating how Jupiter determines to send a notification whenever there is
a failed endpoint. Previously Jupiter used a "enable_calendar" config
value to determine how ALL endpoints would send out notifications. If
this config value was true then all endpoints would use the Infoblox
calendar to determine if a notification was sent, and if it was false
then ALL endpoints would not use the Infoblox calendar to determine if a
notification was sent. This has been changed to a more specified setup
where each endpoint has a "calendar" variable to determine whether the
endpoint uses the Infoblox calendar to determine if a notification is
sent.

Changes include:
- Fixing the config files to use the correct bucket based on the
  environment.
- Changing "summarize" method to include information about the failed
  endpoints. This method has two new booleans indicating if there were
  failed endpoints that use the calendar and if there were failed
  endpoints that do not use the calendar. This method also creates a
  specific message for the failed endpoints that do not use the calendar.
  This message was created because it is needed when it is not
  notification time and there were failed endpoints that use and do not
  use the calendar ( we need to return a message only containing
  information about the failed endpoints that do not use the calendar).
- Changing "check_skip_notification" to use the new variables created in
  summarize. This method was also changed to use the following
  notification configuration: failed endpoints that do not use the
  calendar are always notified and failed endpoints that DO use the
  calendar send a notification if it is the correct time. (The comments
  in the code explain the scenarios)