Created Makefile commands for deploying endpoints
Two makefile commands "deploy-endpoints-test" and
"deploy-endpoints-prod" have been created to update the endpoints.json
file in S3. This S3 file is used by Jupiter and is supposed to be kept
up to date. Previously, to update the endpoints on S3 you had to use a method in Jupiter
that had be to be ran manually. Using the makefile for this action makes
more sense and is simpler.