Added staticmethods as helper functions
These are the static methods are used in 'process_checking' method for detail checking. This new commit includes changes of removing a few methods, single file existence checking and generate pages. The reason to remove these methods because it can be replaced by some existence methods or simpler in process checking function.