Updated Rorschach's load config method and Result
This method originally returned either a boolean upon success or a tuple
upon exception. The return type needed to be the same for both. The
monitor method was also updated to reflect the returned tuple, and the
logic was simplfied. The original logic would break because None type
is not iterable, so it can't be used with the 'in' keyword.

The Result the was created when the config failed to load, tried to
access s3_targets as a list when it is None, so the traceback was passed
to the method instead and logged in the exception message.

The s3 targets config files are loaded based on ENVIRONMENT setting.