Refactored Rorschach's process checking function
This function was overly complicated and only needs to keep track of
which checks fail. This refactor implemented a new dict structure, and
it also provides more metadata for better messaging in the future. This
method still needs to be broken in the future, and it still has a lot of
redundancy that can be reduced with another iteration of work.

Some naming schemes were changed for clarity purposes, and a lot of
comments were added.