Added attribute to keep track of success targets
The original design for checking the various targets is not super
friendly with creating summaries and Results. There was no way to easily
find the targets that succeeded, therefore this success attribute was
added to the processed_targets dict structure.

There is also the addition of an exception list, so in the case of
multiple exceptions, all of them can be recorded.