Refactored summarize method for new dict structure
The processed targets dict structure was altered in a previous commit,
so these changes are in response to that. A few messages were added as
well as some slight structing to the longer email details. These will
need to be tested to determine if they are good enough, but they will
have to be refined in the future.