Fixed Comedian message: floating point precision
Previously the Comedian was having a problem by showing the current
threshold with unnecessary decimals like "56.000000001". This problem
existed previously with the quota_used, but we fixed it using a round().
To make things consistent, I removed that round() function call, and
both percentages displayed in the notifications (threshold percentage,
and used percentage) now will be formatted to 2 decimal places by using
the following "{:.2f}" formatting method.