Refactored utils/s3.py check_bucket method
This method was refactored to be more logical and have better return
values. It now will correctly check if a bucket exists in S3 and return
a boolean and a possible exception. If the s3 bucket check is
successfull it will return True, None. If the s3 bucket check is
unsuccessful and the bucket doesn't exist then False, None is
returned. If an exception is encountered then None, exception is
returned. The unit tests for s3.py have been updated.