Added single file size check.
Previously, single files weren't able to be checked for their size. A
new check has been added to the check_single_file() method to check the
file size if the config value is set in the s3_targets file. This config
value name is the same one used for multiple files. A new message for
the single file check has been added to the messages.py file.