Refactored "_create_summary" in Rorschach
This method has been renamed to "_create_summary_parameters" to match the
other Watchmen. This method has also been significantly refactored. It
no longer builds the messages for each possible type of failure or
exception. Instead, these messages are created when the failure or
exception is found in the process_checking method. This method goes
through all the targets and creates a dictionary of all the result
parameters for all the S3 targets checked.