Refactored "_create_result" in Rorschach
The previous "_create_result" method in Rorschach was too long and had
bad logic so it has been refactored to work the return values from
"create_summary_parameters". This method has been simplified to just
take the parameters from each dictionary in summary_parameters and
create a Result object. At the end of the method a generic Result object
which contains details for all of the target checks is created.