Refactored _check_single_file_existence()
This method in Rorschach has been refactored to have new parameters and
return variables. The s3_key that is used to check if the file exists will be
generated in "_check_single_file" method, which calls this method. The
reason this change was implemented is because the s3_key was being
called multiple times in different methods, which was redundant.

The return variables have been modified as well since we no longer need
to return the generated s3_key since it was passed in.