Refactored _check_file_prefix_suffix in Rorschach
This method has been refactored to only check the suffix. The contents,
which contains all of the files, are generated using the prefix in the
s3_targets config file. So all files in the contents are guaranteed to
have the correct prefix. The message for the suffix check has been
updated to include the exact filename that was found that didn't match
the required suffix.

This check has also been changed to be an opt-in check. The only default
checks are the bucket and empty file checks.