Refactored _generate_contents in Rorschach
This method has been refactored to return a dictionary and traceback
instead of 5 return variables. The dictionary returned contains the
"contents" containing all of the files and their information, the
"count" containing the length of the contents, and "s3_prefix" which has
the full s3 prefix where all of the files from the contents were found.