Removed check_most_recent_file from Rorschach
This variable name and the _check_most_recent_file() method associated
with it are unnecessary. The offset variable is the only required
variable to indicate that file(s) should be checked starting from a
certain amount of time ago. For example, if there is a single daily file
that is checked with an offset=2, then the file will be checked for
existence from 2 days ago. The generate_pages() method used in
generate_contents performs basically the same action as
check_most_recent_file(), so check_most_recent_file() can be removed
safely.