Used cloudformation.yaml as the basis for cf_atg.yaml for our atg environment.
The description was updated to describe the new file function. The account variable was added to be able to update the makefile with an associated AWS account. The original watchmen were kept along with a rorschach watchmen addition. The subnet information was no longer needed, as well as the VPC. The BuildFiles and buckets were updated to reflect which buckets were available in the atg environment.