Changed result_svc to use new python notifiers
Also added a default test SNS arn if testing locally in the test env.
This is because the test SNS topics will not be available locally if
they are store in lambda environment variables. The credentials are from
the boto3 session so as long as there is a WatchmenTest sns topic in the
config file and in the account you are testing from, it will work