Updated Prometheus LZ to Hourly from Minutely
Due to the influx of emails, I decided to change Prometheus LZ to hourly to avoid overflowing people's inboxes with notifications. I did not include a max items but from my testing, it appears there are roughly 500-550 files that are being checked. The minimum file check is 300 files to create a buffer for any time with lower data activity.