Trim contents to the minute rather than the second
Whenever we're looking back a specific offset we don't want to take
seconds into account. This can make it so we are effectively checking a
range that is a few seconds later than what we want to check. In some
cases this means we will miss out on files that would normally be in the
range. This leads to false positive alerts like the ones in TASPT-15