Rorschach time_offset spans multiple days
This commit adds the ability for Rorschach to have a time_offset that
spans multiple days when checking an S3 prefix that is variable by date.

Example: Rorschach runs at 1am UTC with a time_offset of 4 hours for a
prefix of `mypath/year=%0Y/month=%0m/day=%0d/` (on the 4th day of the
month). Before this change we would only check files in the last 3 hours
of the 3rd day of the month. We should also be checking files in the 1st
hour of the 4th day but since the prefix was different for that we were
missing them.
With this change we should now return 2 prefixes to check and load files
from each of them.