Changed Pagination Max File Limit
The Threatwave target is running into issues where false positive
alerts are being sent because the S3 file objects grabbed are less than
all that exist. This leads to one off alerts because sometimes there is
not files that exist within its check threshold in the current
limit but those files do exist on S3.

NOTE: This change is not a guranteed fix. It is just a best effort
for now. There will be a V1 story to look into partitioning the
Threatwave data down to the hour instead of day.