Updated path for Poseidon DNS Data - CZ
After recent changes to the DNS pipeline, we are no longer using the dns-logs/customer prefix and recently updated it to dns-logs/customer-history. At the end of day, the DNS CZ data is suppose to condense and move from customer-current to customer-history and all refinement steps are now reading from customer-history.