Results are saved no matter what
The way this was written previosuly meant that no results would be
saved if an error ocurred trimming them down to not be redundant.
Because this is the first time we are using this storage service and
saving Results to S3, it is better to see what errors or mistakes are
being saved in order to help fix them. We can also make our queries to
S3 smarter in order to avoid these bad results if they occur.