Create New Prefix Offset Attribute
This attribute allows a target to customize the time offset and the
prefix offset separately. This is important for FedPhish because it
inlcudes an hour directory that all the other targets do not.

FedPhish is a tricky target because the timing is absolutely crucial,
and if the lambda doesn't run at exactly 9:00, for example, it may send
a false alert.

Adding the prefix attribute allows the FedPhish target look in the most
recent hour path AND grab files that have been created in the last hour
without worrying what time the lambda runs.