Updated Moto, Boto3, and Botocore
Boto3 and BotocoreI have been updated to the minimum versions that work
with the latest version of Moto. The version of Moto that we were using
ran into dependency error on networkx. This was causing many unit tests
to fail, and the coverage was around 25%. Now, unit tests run again!