Updated deploy-cf.sh to use tag JSON files
The Watchmen project is setup differently from almost all other
cyberapps projects in that it doesn't not load a parameters file. It
defines the parameters at runtime. This solution is not great, and it
shows that we need to rethink the Watchmen CF or wait until it goes to
Atlas.